fixed fade out on icy explosion

+++ The Fester Forest/src/main/java/com/gmail/thelilchicken01/tff/particle/IcyExplosionParticle.java
@@ -37,7 +37,7 @@ public class IcyExplosionParticle extends TextureSheetParticle {
 	}
 
 	private void fadeOut() {
-		this.alpha = (-(1/lifetime) * age + 1);
+		this.alpha = (-(1/(float)lifetime) * age + 1);
 	}
 	
 	@Override
